chore():fixed minor bug & add readme

+++ database/migrations/2023_04_24_091943_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->decimal("product_price", 6,2);
             $table->boolean("product_published");
             $table->integer("product_state");
-            $table->integer("product_referenceId");
+            $table->bigInteger("product_referenceId");
         });
     }
